requirements.txt
numpy
matplotlib
keras
pillow
opencv-python
streamlit
scikit-learn